move routes to controllers
