+++ back-end/src/controllers/story.ts
@@ -1,22 +1,22 @@
 import { Router } from 'express';
 
-import { Category, Game, GamePhase, StoryEntry } from './.generated/prisma';
+import { Category, Game, GamePhase, StoryEntry } from '../.generated/prisma';
 import {
   StoryArchiveResBody as ArchiveRes,
   StoryReqBody,
   StoryResBody,
   UserDto,
-} from './domain/types.js';
-import { joinPhase, loadStory, loadUser } from './middleware.js';
-import prisma from './prisma';
-import { WAIT, punctRegex, quoteRegex } from './utils/constants.js';
-import { ReqHandler as Handler, ReqBody as ReqBody } from './utils/types.js';
+} from '../domain/types.js';
+import { joinPhase, loadStory, loadUser } from '../middleware.js';
+import prisma from '../prisma';
+import { WAIT, punctRegex, quoteRegex } from '../utils/constants.js';
+import { ReqHandler as Handler, ReqBody as ReqBody } from '../utils/types.js';
 import {
   getSuggestion,
   lowerFirst,
   randomNumber,
   upperFirst,
-} from './utils/utils.js';
+} from '../utils/utils.js';
 
 const fillers = ['', '(Man) ', '(Man) and (Woman) ', '', '', ''];
 const prefixes = ['', 'and ', 'were ', 'He said, "', 'She said, "', 'So they '];
@@ -227,8 +227,9 @@ const getArchive: Handler<ReqBody, ArchiveRes> = async (req, res, next) => {
   }
 };
 
-export const router = Router();
+const router = Router();
 router.get('/:id', getArchive);
 router.use(loadUser, loadStory);
 router.get('/', joinPhase, getGame);
 router.put('/', saveEntry);
+export default router;
