+++ back-end/src/middleware.ts
@@ -87,21 +87,3 @@ export const loadStory: Handler = async (req, res, next) => {
   if (req.game.type !== GameType.STORY) return res.sendStatus(400);
   return next();
 };
-
-/**
- * Middleware that logs incoming http requests
- * @param req
- * @param res
- * @param next
- */
-export const accessLogger: Handler = async (req, res, next) => {
-  res.on('finish', () => {
-    if (req.originalUrl.endsWith('/health')) return;
-    console.log(
-      `${new Date().toISOString()} ${req.method} ${req.originalUrl} ${
-        res.statusCode
-      }`
-    );
-  });
-  return next();
-};
