+++ back-end/src/server.ts
@@ -3,11 +3,11 @@ import cookieSession from 'cookie-session';
 import dotenv from 'dotenv';
 import express from 'express';
 
-import { router as gameRoutes } from './games.js';
-import { accessLogger } from './middleware.js';
-import { router as namesRoutes } from './names.js';
-import { router as storyRoutes } from './story.js';
-import { router as userRoutes } from './users.js';
+import gameController from './controllers/games.js';
+import namesController from './controllers/names.js';
+import storyController from './controllers/story.js';
+import userController from './controllers/users.js';
+import { accessLogger } from './utils/accessLogger.js';
 import { notFoundHandler, serverErrorHandler } from './utils/errorHandlers.js';
 
 dotenv.config();
@@ -24,17 +24,16 @@ app.use(
 );
 app.use(accessLogger);
 
-app.use('/api/game', gameRoutes);
-app.use('/api/user', userRoutes);
-app.use('/api/story', storyRoutes);
-app.use('/api/names', namesRoutes);
+app.use('/api/game', gameController);
+app.use('/api/user', userController);
+app.use('/api/story', storyController);
+app.use('/api/names', namesController);
 
 app.get('/health', async (req, res) => {
   res.sendStatus(200);
 });
 
-app.use(notFoundHandler);
-app.use(serverErrorHandler);
+app.use(notFoundHandler, serverErrorHandler);
 
 const runPort = process.env.NODE_PORT || 3000;
 app.listen(runPort, () => console.info(`Server listening on port ${runPort}!`));
